feat: SAP Fiori-style redesign + critical backend fixes (v2.23.0)
Backend
- Add ::float8 casts to all dynamic-SQL NUMERIC selects (rooms, rates,
  room_types, bookings, inventory, reports). Fixes blank prices, rates
  defaulting to 1.0, and zero revenue on every list/report page.
- Fix invoice 500 on walk-in check-ins: COALESCE LEFT-JOIN fields and
  use sqlx force-NOT-NULL hints (new_invoice.rs).
- Add missing inventory mutation/listing endpoints:
  GET  /api/new/inventory/rooms
  POST /api/new/inventory/rooms/:id/check
  POST /api/new/inventory/rooms/:id/replenish
  POST /api/new/inventory/adjustments
- Reshape get_room_inventory response to match RoomInventoryChecklist.
- Add branch=hfhotel|hfville|all filter to /new/* list handlers
  (Ville returns empty until that DB exists; documented inline).
- Regenerate .sqlx/ cache.

Deploy hardening
- migrate.sh: psql -v ON_ERROR_STOP=1 + \\set ON_ERROR_STOP on inside
  the txn so failed migrations are no longer silently marked applied.
- docker-compose: web depends_on backend with condition: service_healthy;
  backend gets a curl /api/mode healthcheck (Dockerfile installs curl).

Design system (SAP Fiori Compact, blood-red theme)
- Sarabun font (Thai + Latin), <html lang=\"th\">, body #F5F6F7, 13px base.
- Blood-red palette (brand-500 #8B0000), all radii squashed to 2px.
- Rewrite Button, Input, Card, Badge, StatCard, DataTable.
- Sidebar: Fiori-style active state (brand-50 bg, brand-500 left border),
  drop Hotel icon and red NEW pill, denser rows.
- NewNavbar: thin 40px top bar with breadcrumb.
- Dashboard tiles flattened, modal de-shadowed, page headers slimmed.

Versions
- package.json 2.22.1 -> 2.23.0
- hotel-backend/Cargo.toml 2.7.0 -> 2.8.0

Tests: 655/655 frontend, 9/9 backend suites, prod build clean.

+++ hotel-backend/src/routes/new_reports.rs
@@ -138,7 +138,7 @@ pub async fn get_revenue(
         )
         SELECT
             period,
-            COALESCE(SUM(revenue), 0) as total_revenue,
+            COALESCE(SUM(revenue), 0)::float8 as total_revenue,
             COUNT(*)::int as booking_count
         FROM DateRange
         GROUP BY period
@@ -239,7 +239,7 @@ pub async fn get_occupancy(
                      ELSE ci.cin_checkin_time::date END)
             ), 0) as occupied_nights,
             COALESCE(SUM(COALESCE(ci.cin_total_amount,
-                ci.cin_rate_per_night * (COALESCE(ci.cin_checkout_time, ci.cin_expected_checkout)::date - ci.cin_checkin_time::date))), 0) as total_revenue,
+                ci.cin_rate_per_night * (COALESCE(ci.cin_checkout_time, ci.cin_expected_checkout)::date - ci.cin_checkin_time::date))), 0)::float8 as total_revenue,
             COUNT(*)::int as checkin_count,
             COALESCE(AVG((COALESCE(ci.cin_checkout_time, ci.cin_expected_checkout)::date - ci.cin_checkin_time::date)::float8), 0) as avg_stay
         FROM ht_checkins ci
@@ -321,7 +321,7 @@ pub async fn get_revenue_by_room_type(
         SELECT
             COALESCE(rt.type_name, 'Unknown') as room_type,
             COALESCE(SUM(COALESCE(ci.cin_total_amount,
-                ci.cin_rate_per_night * (COALESCE(ci.cin_checkout_time, ci.cin_expected_checkout)::date - ci.cin_checkin_time::date))), 0) as revenue
+                ci.cin_rate_per_night * (COALESCE(ci.cin_checkout_time, ci.cin_expected_checkout)::date - ci.cin_checkin_time::date))), 0)::float8 as revenue
         FROM ht_checkins ci
         LEFT JOIN ht_rooms_new r ON ci.cin_room_id = r.room_id
         LEFT JOIN ht_room_types rt ON r.room_type_id = rt.type_id
